refactored auth service tests
