+++ server/__tests__/auth/auth.services.test.ts
@@ -3,47 +3,24 @@ import * as bcrypt from 'bcrypt'
 import { createMockContext, MockContext } from "../../src/context/mockContext"
 import authServices from '../../src/auth/auth.services'
 import authJwt from "../../src/auth/auth.jwt"
+import mockData from "../mockData"
 
 let mockCtx: MockContext
 beforeEach(() => {
   mockCtx = createMockContext()
 })
 
-// Mock Data
 const currentDate = new Date()
 const refreshToken = 'refresh-token'
 const hashedToken = 'hashed-refresh-token'
 
-const mockRefreshToken = {
-  id: '1',
-  createdAt: currentDate,
-  updatedAt: currentDate,
-  hashedToken,
-  userId: '1',
-  revoked: false,
-  expireAt: new Date(Date.now() + 1000 * 60 * 60 * 24 * 30),
-}
-
-const mockUser = {
-  id: "1",
-  email: 'test@gmail.com',
-  password: '12345678',
-  firstName: 'test',
-  lastName: 'user',
-  accountType: UserRole.Professor,
-  socketId: null,
-  createdAt: currentDate,
-  updatedAt: currentDate
-}
-
-// Tests
 test('createUserByEmailAndPassword creates a new user with hashed password', async () => {
-  const hashedPassword = await bcrypt.hash(mockUser.password, 12)
+  const hashedPassword = await bcrypt.hash(mockData.mockUser.password, 12)
   mockCtx.prisma.user.create.mockResolvedValue({
-    ...mockUser,
+    ...mockData.mockUser,
     password: hashedPassword,
   })
-  const result = await authServices.createUserByEmailAndPassword(mockUser, mockCtx)
+  const result = await authServices.createUserByEmailAndPassword(mockData.mockUser, mockCtx)
   expect(await bcrypt.compare('12345678', result.password)).toBe(true)
   expect(result).toMatchObject({
     id: "1",
@@ -57,41 +34,41 @@ test('createUserByEmailAndPassword creates a new user with hashed password', asy
 })
 
 test('findUserById returns the user object with the matching Id', async () => {
-  mockCtx.prisma.user.findUnique.mockResolvedValue(mockUser)
+  mockCtx.prisma.user.findUnique.mockResolvedValue(mockData.mockUser)
   const result = await authServices.findUserById("1", mockCtx)
-  expect(result).toMatchObject(mockUser)
+  expect(result).toMatchObject(mockData.mockUser)
 })
 
 test('findUserByEmail returns the user object with the matching email', async () => {
-  mockCtx.prisma.user.findUnique.mockResolvedValue(mockUser)
+  mockCtx.prisma.user.findUnique.mockResolvedValue(mockData.mockUser)
   const result = await authServices.findUserById("test@gmail.com", mockCtx)
-  expect(result).toMatchObject(mockUser)
+  expect(result).toMatchObject(mockData.mockUser)
 })
 
 test('addRefreshTokenToWhiteList adds a new refresh token to the whitelist', async () => {
-  mockCtx.prisma.refreshToken.create.mockResolvedValue(mockRefreshToken)
+  mockCtx.prisma.refreshToken.create.mockResolvedValue(mockData.mockRefreshToken)
   jest.spyOn(authJwt, 'hashToken').mockReturnValue(hashedToken)
-  const result = await authServices.addRefreshTokenToWhiteList(refreshToken, mockUser.id, mockCtx)
-  expect(result).toMatchObject(mockRefreshToken)
+  const result = await authServices.addRefreshTokenToWhiteList(refreshToken, mockData.mockUser.id, mockCtx)
+  expect(result).toMatchObject(mockData.mockRefreshToken)
 })
 
 test('findRefreshToken finds a refresh token by its token string', async () => {
-  mockCtx.prisma.refreshToken.findUnique.mockResolvedValue(mockRefreshToken)
+  mockCtx.prisma.refreshToken.findUnique.mockResolvedValue(mockData.mockRefreshToken)
   jest.spyOn(authJwt, 'hashToken').mockReturnValue(hashedToken)
   const result = await authServices.findRefreshToken(refreshToken, mockCtx)
-  expect(result).toMatchObject(mockRefreshToken)
+  expect(result).toMatchObject(mockData.mockRefreshToken)
 })
 
 test('deleteRefreshTokenById deletes a refresh token by its ID', async () => {
-  const deletedToken = { ...mockRefreshToken, revoked: true }
+  const deletedToken = { ...mockData.mockRefreshToken, revoked: true }
   mockCtx.prisma.refreshToken.update.mockResolvedValue(deletedToken)
-  const result = await authServices.deleteRefreshTokenById(mockRefreshToken.id, mockCtx)
+  const result = await authServices.deleteRefreshTokenById(mockData.mockRefreshToken.id, mockCtx)
   expect(result).toMatchObject(deletedToken)
 })
 
 test('revokeTokens revokes all refresh tokens for a user', async () => {
   mockCtx.prisma.refreshToken.updateMany.mockResolvedValue({ count: 2 })
-  const result = await authServices.revokeTokens(mockUser.id, mockCtx)
+  const result = await authServices.revokeTokens(mockData.mockUser.id, mockCtx)
   expect(result).toEqual({ count: 2 })
   expect(result.count).toBe(2)
 })
